+++ rs/pocket_ic_server/src/pocket_ic.rs
@@ -19,14 +19,13 @@ use ic_types::messages::CertificateDelegation;
 use ic_types::{CanisterId, PrincipalId, SubnetId};
 use itertools::Itertools;
 use pocket_ic::common::rest::{
-    self, BinaryBlob, BlobCompression, RawAddCycles, RawCanisterCall, RawEffectivePrincipal,
-    RawSetStableMemory, SubnetConfigSet, SubnetKind, Topology,
+    self, BinaryBlob, BlobCompression, ExtendedSubnetConfigSet, RawAddCycles, RawCanisterCall,
+    RawEffectivePrincipal, RawSetStableMemory, SubnetKind, SubnetSpec, Topology,
 };
 use rand::rngs::StdRng;
 use rand::Rng;
 use rand::SeedableRng;
 use serde::{Deserialize, Serialize};
-use std::iter::repeat;
 use std::str::FromStr;
 use std::{
     collections::{BTreeMap, HashMap},
@@ -52,11 +51,18 @@ pub struct PocketIc {
 }
 
 impl PocketIc {
-    pub fn new(runtime: Arc<Runtime>, subnet_configs: SubnetConfigSet) -> Self {
+    pub fn new(runtime: Arc<Runtime>, subnet_configs: ExtendedSubnetConfigSet) -> Self {
         let fixed_range_subnets = subnet_configs.get_named();
         let flexible_subnets = {
-            let sys = repeat((SubnetKind::System, None)).take(subnet_configs.system);
-            let app = repeat((SubnetKind::Application, None)).take(subnet_configs.application);
+            // note that for these, the subnet ids are currently ignored.
+            let sys = subnet_configs
+                .system
+                .iter()
+                .map(|spec| (SubnetKind::System, spec.get_path()));
+            let app = subnet_configs
+                .application
+                .iter()
+                .map(|spec| (SubnetKind::Application, spec.get_path()));
             sys.chain(app)
         };
 
@@ -65,9 +71,10 @@ impl PocketIc {
         let mut subnet_ids = vec![];
         let mut routing_table = RoutingTable::new();
 
-        let mut nns_subnet_id: Option<SubnetId> = subnet_configs
-            .nns_subnet_id
-            .map(|raw_nns_subnet_id| SubnetId::new(PrincipalId(raw_nns_subnet_id.into())));
+        let mut nns_subnet_id = subnet_configs.nns.and_then(|x| {
+            x.get_subnet_id()
+                .map(|y| SubnetId::new(PrincipalId(y.into())))
+        });
 
         let mut subnet_counter = 0_u64;
         let mut apply_subnet_counter = move || -> u64 {
@@ -275,8 +282,8 @@ impl Default for PocketIc {
     fn default() -> Self {
         Self::new(
             Runtime::new().unwrap().into(),
-            SubnetConfigSet {
-                application: 1,
+            ExtendedSubnetConfigSet {
+                application: vec![SubnetSpec::New],
                 ..Default::default()
             },
         )
@@ -1131,8 +1138,8 @@ mod tests {
     fn new_pic_counter_installed_system_subnet() -> (PocketIc, CanisterId) {
         let mut pic = PocketIc::new(
             Runtime::new().unwrap().into(),
-            SubnetConfigSet {
-                ii: true,
+            ExtendedSubnetConfigSet {
+                ii: Some(SubnetSpec::New),
                 ..Default::default()
             },
         );